% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlog.R
\name{dlog}
\alias{dlog}
\title{dlog}
\usage{
dlog(x, convert = TRUE)
}
\arguments{
\item{x}{Time series or vector to be converted to a \code{ts} object.}

\item{convert}{(Optional) Controls the automatia conversion to index before passing \code{log()}. Defaults to \code{TRUE}.}
}
\value{
Time series is the percentage variation with the first position as \code{0}.
}
\description{
Generates series of the percentage variation of the series data passed as parameter.
}
\details{
It is implicit that data is index by time (time series). If observations are negative or zero, the \code{convert=TRUE} option allows the automatic transformation to an index before passing the \code{log()}.
}
\examples{
x <- ts( rnorm(10,mean=100) )
dlx <- dlog(x)

x1 <- ts( seq(-.3,.6,.1) )
dlog( x1, convert=FALSE ) # causes NA when log(x) for x<=0
}
\seealso{
\code{\link{toIndex}}
}

