% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genEVtime.R
\name{genEVtime}
\alias{genEVtime}
\title{genEVtime}
\usage{
genEVtime(start, end = NULL, qobs = NULL)
}
\arguments{
\item{start}{String with the starting date formatted as \code{"2000M12"} or \code{"2002Q4"}.}

\item{end}{String with the ending date, uses the same format as \code{start}. Optional if \code{qobs} specified. Defaults to NULL.}

\item{qobs}{Number of observations. Optional if \code{end} specified. Defaults to NULL.}
}
\description{
Function to generate an Eviews-formatted vector of strings with time references.
}
\details{
Eviews formats dates of regularly spaced time series as YYYYPSS where YYYY is the year, P is the period (Month,Quarter) and SS is the subperiod (eg.1 to 12 for months)
}
\examples{
genEVtime("1973Y1","1975Y1") # yeap, must explicit "Y1"
genEVtime("1947q3","1948Q02") # note the 0 before 2 is optional
genEVtime("1947M1",qobs=6) #

}

