% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macror.R
\docType{package}
\name{macror-package}
\alias{macror}
\alias{macror-package}
\title{Macroeconomic analysis in R: Set of functions to work with time series \code{ts} and \code{zoo} objects.}
\description{
Package \code{macror} speeds up tasks related to time series (ts) and zoo objects: 
\code{\link{dlog}}, \code{\link{urTable}}, \code{\link{genEVtime}}, \code{\link{tsVarToIndex}}, \code{\link{tsNumToIndex}}, \code{\link{tsComplete}}, 
\code{\link{tsDummy}} and \code{\link{tsMapOutliers}}.
}

