% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOutliers.R
\name{removeOutliers}
\alias{removeOutliers}
\title{removeOutliers}
\usage{
removeOutliers(x, range = c(0.01, 0.99), na.rm = TRUE, ...)
}
\arguments{
\item{x}{Series to be filtered}

\item{na.rm}{Control NA removal with TRUE (to remove NAs) or FALSE (to keep them).}

\item{ok_range}{Quantile range of accepted values}
}
\description{
Function to remove "outlier" observations as defined by quantile range of acceptable values.
}
\details{
Source: \url{http://stackoverflow.com/questions/4787332/how-to-remove-outliers-from-a-dataset}
}
\examples{

set.seed(123)
x <- rnorm(100)
x <- c(-10, x, 10)
y <- removeOutliers(x)

}

