% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toIndex.R
\name{toIndex}
\alias{toIndex}
\title{toIndex}
\usage{
toIndex(x, base = 1)
}
\arguments{
\item{x}{Time series to be converted to index.}

\item{base}{(optional) is the index of the observaton to be used as reference.}
}
\description{
Convert series to index based 100 corresponding to the reference observation given in optional parameter \code{base}.
}
\details{

}
\examples{
Inflation <- ts( rnorm(13,mean=.2,sd=.1), start=c(2007,12), freq=12 )
Cpi <- toIndex(Inflation)

}
\seealso{
\code{\link{dlog}}
}

