% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsComplete.R
\name{tsComplete}
\alias{tsComplete}
\title{Completes a series with data from anothe series or by interpolating and/or repeating.}
\usage{
tsComplete(x, y, fill = "fill")
}
\arguments{
\item{x}{main data series (either \code{ts} or \code{zoo}) to be completed for missing observations.}

\item{y}{secondary series (either \code{ts} or \code{zoo}) to use only when missing in the main series.}

\item{fill}{method to treat NA's after combining \code{x} and \code{y}, can be "na","fill","interpolate" or "repeat"}
}
\value{
Returns either a \code{ts} (default) or \code{zoo} object.
}
\description{
Function to complete a time-indexed series (either \code{ts} or \code{zoo}) by combining it with another series to avoid NAs. The function has a third argument "fill" to chose whether to leave NA, interpolate or repeate the last value when both vectors are missing.
}
\details{
Supports both \code{ts} and \code{zoo}. Converts to \code{zoo} if one of the parameters is \code{zoo}.
}
\examples{
a  <- c(NA, 1, 2,NA,NA, 5,NA)
b  <- c(   NA,12,NA,13,NA,15,16)
a2 <- ts(a,start=c(2000,1),freq=4)
b2 <- ts(b,start=c(2000,2),freq=4)
b3 <- zoo::as.zoo(b2)
aa  <- tsComplete(a2)
ab1 <- tsComplete(a2,b2)
ab2 <- tsComplete(a2,b2,fill="na")
ab3 <- tsComplete(a2,b3)
ab4 <- tsComplete(a2,b3,fill="fill")
}

