% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsNumToIndex.R
\name{tsNumToIndex}
\alias{tsNumToIndex}
\title{Converts a numeric series to an index using the first observation as reference.}
\usage{
tsNumToIndex(x, base = 100, reference = 1)
}
\arguments{
\item{x}{Time-indexed data to convert, must be a ts or zoo object.}

\item{base}{(Optional) Default is 100.}

\item{reference}{(Optional) Observation to be used as proportions are calculated.}
}
\value{
Time series (ts or zoo) with the index.
}
\description{
Converts a numeric series to an index using the first observation as reference.
}
\details{
If using a monetary series be sure to deflate it before indexing.
}
\examples{
# Production averages 1254 units per month  
production <- ts( rnorm(30*12,mean=1254,sd=425),start=c(1980,1),freq=12)
ip  <- tsNumToIndex( production ) # industrial production index 
ip2 <- tsNumToIndex( production, reference = c(1992,6) ) # rebase for June 1992
}
\seealso{
\code{\link{tsVarToIndex}}
}

