% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urTable.R
\name{urTable}
\alias{urTable}
\title{Generates a table with Unit Roor test results}
\usage{
urTable(df, tests = c("adf", "pp", "kpss"), order = 1, file = NULL,
  format = "csv")
}
\arguments{
\item{df}{Dataframe with series to test}

\item{order}{Maximum order of integration, I(?), to test starting from I(0). It can be\code{c(0,1,2)}.}

\item{file}{(Optional) Filename to be generated according to \code{format}.}

\item{format}{(Optional) Format of the output table, can be \code{c("csv","latex")}. Defaults to \code{"txt"}.}

\item{structure}{List of models to test, can be \code{c("c","ct","nc")}: \code{"c"} assumes model with intercept,"ct" assumes intercept and deterministic trend,"nc" assumes no intercept and no trend.}
}
\value{
A dataframe with the test results and a file (optional).
}
\description{
Returns a datatable and generates a .csv or .tex file with Unit Root test results.
}
\details{
Supports ADF, PP and KPSS unit root tests.
}
\examples{
data <- data.frame( USeconomic ) #from tseries 
dft1 <- urTable( data, tests=c("adf","pp","kpss"), file="US.csv" )
dft2 <- urTable( data, tests=c("pp","kpss"), order=2, file="US.tex", format="latex" )

}

