% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlog.R
\name{dlog}
\alias{dlog}
\title{Calculates the percentage variation of a time series.}
\usage{
dlog(data, lag = 1)
}
\arguments{
\item{data}{Time series or vector to be converted to a \code{ts} object.}

\item{lag}{(Optional) Number of lags to differentiate with \code{diff()}. Defaults to \code{1}.}
}
\value{
Time series of the percentage variation with the first position(s) as \code{NA}.
}
\description{
Generates series of the percentage variation of the series data passed as parameter.
}
\details{
Data must be time indexed, either as \code{ts} or \code{zoo}. If observation(s) are negative or zero, the algorithm changes to the traditional percentage variation.
}
\examples{
CPI <- ts( rnorm(10,mean=100) )
Inflation <- dlCPI <- dlog( CPI )  # choose your naming convention!

x <- ts( seq(-.3,.6,.1) )
dlx <- dlog( x ) # switch algorithm to avoid generating NAs 
dl2x <- dlog( x, lag=2 ) 
}
\seealso{
\code{\link{pVarToIndex}}
}

