% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numToIndex.R
\name{numToIndex}
\alias{numToIndex}
\title{numToIndex}
\usage{
numToIndex(x, reference = 1, base = 100)
}
\arguments{
\item{x}{Time-indexed data to convert, must be a ts or zoo object.}

\item{reference}{(Optional) Observation to be used as proportions are calculated.}

\item{base}{(Optional) Defaults to 100.}
}
\value{
Time series (ts or zoo) with the index.
}
\description{
Converts a numeric series to an index using the first observation as reference.
}
\details{
If using a monetary series be sure to deflate it before indexing.
}
\examples{
# Production averages 1254 units per month  
production <- ts( rnorm(30*12,mean=1254,sd=425),start=c(1980,1),freq=12)
ip <- numToIndex( production ) # industrial production index 
}
\seealso{
\code{\link{pVarToIndex}}
}

