% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pVarToIndex.R
\name{pVarToIndex}
\alias{pVarToIndex}
\title{Converts a percentage variation series into a base 100 index series.}
\usage{
pVarToIndex(x, base = 100, reference = 1)
}
\arguments{
\item{x}{Time series to be converted to index.}

\item{base}{(optional, default is 100).}

\item{reference}{(optional, default is 1) Index of the observaton to be used as reference.}
}
\description{
Compounds a percentage variation series into a 100 based index corresponding to the reference observation.
}
\details{
The index is based in 100. Reference is by default the first observation.
}
\examples{
# Monthly inflation is around 2\%
Inflation <- ts( rnorm(13,mean=2,sd=.5), start=c(2007,12), freq=12 )
CPI_dec07 <- pVarToIndex( Inflation )

}
\seealso{
\code{\link{dlog}}, \code{\link{numToIndex}}
}

