% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsworkflow.R
\docType{package}
\name{tsworkflow-package}
\alias{tsworkflow}
\alias{tsworkflow-package}
\title{Set of functions to work with time series \code{ts} and \code{zoo} objects.}
\description{
Package \code{tsworkflow} speeds up tasks related to time series (ts) and zoo objects: 
\code{\link{dlog}}, \code{\link{genEVtime}}, \code{\link{tsVarToIndex}}, \code{\link{tsNumToIndex}}, \code{\link{tsComplete}}, 
\code{\link{tsDummy}} and \code{\link{tsMapOutliers}}.
}

