% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urTable.R
\name{urTable}
\alias{urTable}
\title{Generates a table with Unit Roor test results}
\usage{
urTable(df, tests = c("adf", "pp", "kpss"), order = 1, file = NULL,
  format = "csv")
}
\arguments{
\item{df}{Dataframe with series to test}

\item{tests}{List of tests to be performed, by default tests are \code{c("adf","pp","kpss")}.}

\item{order}{Maximum order of integration to test, usually noted as I(?). It can be \code{c(0,1,2)}.}

\item{file}{(Optional) Filename to be generated according to \code{format}.}

\item{format}{(Optional) Format of the output table, can be \code{c("csv","latex")}. Defaults to \code{"txt"}.}
}
\value{
A dataframe with the test results and a file (optional).
}
\description{
Returns a datatable and generates a .csv or .tex file with Unit Root test results.
}
\details{
Supports ADF, PP and KPSS unit root tests.
}
\examples{

require(tseries)
data(USeconomic,package="tseries")
data <- data.frame( USeconomic ) 

dft1 <- urTable( data, tests=c("adf","pp","kpss"), file="US.csv" )

dft2 <- urTable( data, tests=c("pp","kpss"), order=2, file="US.tex", format="latex" )

# Even series with missing observations can be tested
bad <- data.frame( var1=rnorm(100), var2=cumsum(rnorm(100)) ) 
bad[seq(95,100),"var1"] <- bad[1,"var2"] <- bad[2,"var2"] <- NA
dft3 <- urTable( bad )
}

